# lib/api-zod/src/generated/types/appSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dashboard Web Personal API
 * OpenAPI spec version: 0.1.0
 */

export interface AppSettings {
  id: number;
  dashboardName: string;
  headerText: string;
  theme: string;
  weatherLocation: string;
  newsCount: number;
  updatedAt: string;
}