# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dashboard Web Personal API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  emailCount24h: number;
  pendingTasks: number;
  fiscalUpdatesCount: number;
  fiscalRequireAction: number;
  travelOffersCount: number;
  newsCount: number;
}