# lib/api-zod/src/generated/types/fiscalUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dashboard Web Personal API
 * OpenAPI spec version: 0.1.0
 */

export interface FiscalUpdate {
  id: number;
  title: string;
  jurisdiction: string;
  category: string;
  organism: string;
  date: string;
  impact: string;
  summary: string;
  requiresAction: boolean;
  isSaved: boolean;
  createdAt: string;
}