# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dashboard Web Personal API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsArticle {
  id: number;
  title: string;
  source: string;
  category: string;
  date: string;
  summary: string;
  url: string;
  /** @nullable */
  imageUrl?: string | null;
}