# lib/api-zod/src/generated/types/shortcut.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dashboard Web Personal API
 * OpenAPI spec version: 0.1.0
 */

export interface Shortcut {
  id: number;
  name: string;
  url: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  userId?: string | null;
  createdAt: string;
}