# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dashboard Web Personal API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: string;
  priority?: string;
  /** @nullable */
  dueDate?: string | null;
}